Move test-utils.
