+++ src/button/__tests__/button.test.tsx
@@ -6,7 +6,6 @@ import { act, fireEvent, render } from '@testing-library/react';
 import Button, { ButtonProps } from '../../../lib/components/button';
 import InternalButton from '../../../lib/components/button/internal';
 import createWrapper, { ButtonWrapper } from '../../../lib/components/test-utils/dom';
-import LiveRegionWrapper from '../../../lib/components/test-utils/selectors/internal/live-region';
 import { buttonRelExpectations, buttonTargetExpectations } from '../../__tests__/target-rel-test-helper';
 import { renderWithSingleTabStopNavigation } from '../../internal/context/__tests__/utils';
 
@@ -64,13 +63,11 @@ describe('Button Component', () => {
   describe.each([true, false])('loadingText property, with href: %s', withHref => {
     test('renders loadingText in a LiveRegion', () => {
       renderButton({ children: 'Button', loading: true, loadingText: 'Loading', href: withHref ? '#' : undefined });
-      expect(createWrapper().findByClassName(LiveRegionWrapper.rootSelector)!.getElement()).toHaveTextContent(
-        'Loading'
-      );
+      expect(createWrapper().findLiveRegion()!.getElement()).toHaveTextContent('Loading');
     });
     test('does not render loadingText if loading is false', () => {
       renderButton({ children: 'Button', loading: false, loadingText: 'Loading', href: withHref ? '#' : undefined });
-      expect(createWrapper().findByClassName(LiveRegionWrapper.rootSelector)).toBeNull();
+      expect(createWrapper().findLiveRegion()).toBeNull();
     });
   });
