+++ src/test-utils/dom/live-region/index.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 import { ComponentWrapper } from '@cloudscape-design/test-utils-core/dom';
 
-import styles from '../../../internal/components/live-region/styles.selectors.js';
+import styles from '../../../live-region/styles.selectors.js';
 
 export default class LiveRegionWrapper extends ComponentWrapper {
   static rootSelector: string = styles.root;
